making SkipList on-disk form: implemented // TODO: (SDB) need UnpinPage of node at appropriate timing.

+++ storage/page/skip_list_page/skip_list_header_page.go
@@ -71,7 +71,7 @@ func NewSkipListStartBlockPage(bpm *buffer.BufferPoolManager, keyType types.Type
 	//startNode.SetForward(make([]*SkipListBlockPage, MAX_FOWARD_LIST_LEN))
 	// set sentinel node at end of list
 	for ii := 0; ii < MAX_FOWARD_LIST_LEN; ii++ {
-		startNode.SetForwardEntry(int32(ii), sentinelNode.GetPageId())
+		startNode.SetForwardEntry(ii, sentinelNode.GetPageId())
 	}
 
 	ret := startNode.GetPageId()
